+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - [feature] Show notifications when a new diff is ready to land
 - [feature] Make whether to show diff notifications configurable
+- [feature] Track usage with google analytics
 - [internal] Improve type information for `.arcrc` file
 - [internal] Always fetch configuration instead of passing it through on activation
 - [internal] Big refactor to make code more readable